==> Notebook_4b_comparison_of_emissions_against_gridded_data.ipynb <==
## Comparison of reservoir emissions in Myanmar calculated explicitly with GeoCARET and ReEmission, and the estimates obtained by extrapolating and binning measured emissions using latitude and climate zone parameterizations

### Author:
##### Tomasz Janus, University of Manchester, tomasz.janus@manchester.ac.uk; tomasz.k.janus@gmail.com
##### Date: 30-10-24, Modified: 23-11-24

### Data Sources:
We use two sets of data representing gridded areal CO$_2$ and CH$_4$ emission estimates:
1) The dataset created by Harrison et al. 2021 [[1](https://doi.org/10.1029/2020GB006888)]
2) The dataset prepared by Soued et al. 2022 [[2](https://doi.org/10.1038/s41561-022-01004-2)]. This dataset shows emission factors vs. climatic zones. It was shared as a table in Supplementary Information and is an extension of emission factors proposed in the 2019 refinement to the 2006 IPCC emission guidelines, that includes additional degassing and ebullition CH$_4$ emission pathways.

### Notes on Data Preprocessing and Algorithms

#### Estimation of reservoir emissions from gridded data

We calculated the net anthropogenic contribution of reservoirs to emissions from the gridded estimates for CO$_2$ and CH$_4$ emissions respectively using equation 1 from Almeida et al. 2019 [[3](https://doi.org/10.1038/s41467-019-12179-5)], presented below:

**Eq. 1** : ${\mathrm{TE}}_{{\mathrm{reservoir}}} = A_{{\mathrm{reservoir}}} \times \left( {{\mathrm{net}}_{{\mathrm{CO}}_{2}} \times F_{{\mathrm{CO}}_{\mathrm{2}}} + {\mathrm{net}}_{{\mathrm{CH}}_{4}} \times F_{{\mathrm{CH}}_{\mathrm{4}}} \times {\mathrm{GWP}}_{{\mathrm{CH}}_{4}}} \right) \times \left( {1 + R_{{\mathrm{downstream}}}} \right)$

where $A_{{\mathrm{reservoir}}}$ is the reservoir area and $F_{{\mathrm{CO}}_{\mathrm{2}}}$ and $F_{{\mathrm{CH}}_{\mathrm{4}}}$ are the areal CO$_2$ and CH$_4$ emissions (fluxes), respectively. ${\mathrm{GWP}}_{{\mathrm{CH}}_{4}}$ is a conversion factor for the global warming potential of CH$_4$ relative to CO$_2$ over the corresponding time horizon (e.g. 20 or 100 years). We have assumed the time horizon of 100 years. The parameters $\mathrm{net}_{\mathrm{CO}_2}$ , $\mathrm{net}_{\mathrm{CH}_4}$ and $R_\mathrm{downstream}$ are introduced by Almeida et al. 2019 [[3](https://doi.org/10.1038/s41467-019-12179-5)] to account only for the true net (anthropogenic) change in GHG emissions associated with reservoir creation from the gross emissions data that account for both the anthropogenic and non-anthropogenic emissions. These coefficients model the discount factors for CO$_2$ and CH$_4$ emissions that determine the fraction of net anthropogenic emissions in the total emissions of CO$_2$ and CH$_4$, respectively, and the ratio of downstream emissions to reservoir-surface emissions. We adopted the original values presented in Almeida et al. 2019 [[3](https://doi.org/10.1038/s41467-019-12179-5)], i.e. $\mathrm{net}_{\mathrm{CO}_2} = 0.25$, $\mathrm{net}_{\mathrm{CH}_4} = 0.90$, and $R_\mathrm{downstream} = 0.17$. Additionally, we fitted new values using linear regression to obtain the best fit between emission predictions from G-res (facilitated via the ReEmission software) and the factors per climatic zone of Soued et al. 

#### IPCC Guidelines

IPCC provides main guidance on reservoir GHG emissions in the Chapter 7, Volume 4 of the 2019 Refinement for a more complete inventory of greenhouse gas emissions from managed lands, ([Lovelock et al. 2019](https://www.ipcc-nggip.iges.or.jp/public/2019rf/pdf/4_Volume4/19R_V4_Ch07_Wetlands.pdf))

### Purpose of the Notebook
The purpose of this notebook is to calculate and visualise differences in estimates of emissions of individual reservoirs as well as total country-wide emissions from reservoirs from the two approaches: (a) direct calculation using G-res emission model for each individual reservoir, and (b) using interpolated values from a limited number of known emissions from reservoirs in different locations across the world.

### References:
* [1] Harrison, J. A., Prairie, Y. T., Mercier-Blais, S., & Soued, C. (2021). Year-2020 global distribution and pathways of reservoir methane and carbon dioxide emissions according to the greenhouse gas from reservoirs (G-res) model. Global Biogeochemical Cycles, 35, e2020GB006888. https://doi.org/10.1029/2020GB006888 
* [2] Soued, C., Harrison, J.A., Mercier-Blais, S. et al. Reservoir CO2 and CH4 emissions and their climate impact over the period 1900–2060. Nat. Geosci. 15, 700–705 (2022). https://doi.org/10.1038/s41561-022-01004-2
* [3] Almeida, R.M., Shi, Q., Gomes-Selman, J.M. et al. Reducing greenhouse gas emissions of Amazon hydropower with strategic dam planning. Nat Commun 10, 4281 (2019). https://doi.org/10.1038/s41467-019-12179-5
* [4] Carlino, A., Schmitt, R., Clark, A. et al. Rethinking energy planning to mitigate the impacts of African hydropower. Nat Sustain 7, 879–890 (2024). https://doi.org/10.1038/s41893-024-01367-x
* [5] Bridget R. Deemer, John A. Harrison, Siyue Li, Jake J. Beaulieu, Tonya DelSontro, Nathan Barros, José F. Bezerra-Neto, Stephen M. Powers, Marco A. dos Santos, J. Arie Vonk, Greenhouse Gas Emissions from Reservoir Water Surfaces: A New Global Synthesis, BioScience, Volume 66, Issue 11, 1 November 2016, Pages 949–964, https://doi.org/10.1093/biosci/biw117
* [6] Catherine Ellen Lovelock, Christopher Evans, Nathan Barros, Yves Prairie, Jukka Alm, David Bastviken, Jake J. Beaulieu, Michelle Garneau, Atle Harby, John Harrison, David Pare, Hanne Lerche Raadal, Bradford Sherman, Chengyi Zhang, Stephen Michael Ogle, 2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories, Volume 4: Agriculture, Forestry and Other Land Use, Chapter 7 Wetlands, https://www.ipcc-nggip.iges.or.jp/public/2019rf/pdf/4_Volume4/19R_V4_Ch07_Wetlands.pdf
* [7] Carly Hansen, Rachel Pilla, Paul Matson, Bailey Skinner, Natalie Griffiths, & Henriette Jager (2023). Variability in modelled reservoir greenhouse gas emissions: comparison of select US hydropower reservoirs against global estimates. Environmental Research Communications, 4(12), 121008. https://iopscience.iop.org/article/10.1088/2515-7620/acae24#ercacae24s6
* [8] Bridget R. Deemer, John A. Harrison, Siyue Li, Jake J. Beaulieu, Tonya DelSontro, Nathan Barros, José F. Bezerra-Neto, Stephen M. Powers, Marco A. dos Santos, J. Arie Vonk, Greenhouse Gas Emissions from Reservoir Water Surfaces: A New Global Synthesis, BioScience, Volume 66, Issue 11, 1 November 2016, Pages 949–964, https://doi.org/10.1093/biosci/biw117
* [9]  Tangi, M., Schmitt, R., Almeida, R., Bossi, S., Flecker, A., Sala, F., & Castelletti, A. (2024). Robust hydropower planning balances energy generation, carbon emissions and sediment connectivity in the Mekong River Basin. Earth's Future, 12, e2023EF003647. https://doi.org/10.1029/2023EF003647 

### Notes on adoption of GHG fluxes in most recent planning studies

1. Almeida et al. 2019 [[3](https://doi.org/10.1038/s41467-019-12179-5)] used the gross flux data from one of the earlier global reservoir emission synthesis by Deemer et al. [[5](https://doi.org/10.1093/biosci/biw117)].
2. Carlino et al. 2024 [[4](https://doi.org/10.1038/s41893-024-01367-x)] analyzed the gridded estimates from Harrison et al. 2021 [[1](https://doi.org/10.1029/2020GB006888)] that are parameterized vs. latitude and the parameterization of Soued et al. 2022 [[2](https://doi.org/10.1038/s41561-022-01004-2)] who parameterized the emissions vs. climatic zone. They processed both datasets using Eq. 1 adopted after Almeida et al. 2019 [[3](https://doi.org/10.1038/s41467-019-12179-5)]. Nevetheless, since only a few points in the dataset of [[1](https://doi.org/10.1029/2020GB006888)] were available for Africa (which was the geographical context of their study), and most of the estimates at these latitudes were from Southeast Asia or Latin America, the authors based their results solely on the greenhouse gas emission estimations of Soued et al. (2022) [[2](https://doi.org/10.1038/s41561-022-01004-2)].

### Notes on selected cited studies

#### * Harrison et al.
The data with the estimated CO$_2$ and CH$_4$ emission fluxes per given latitude has been shared in this [Zenodo link](https://zenodo.org/records/4632428/files/G-res_Output_x_Deg_Latitude_20210423.xlsx?download=1). The authors also present data globally on a 1degx1deg grid, albeit it has not been clearly explained what data each dataset contains. All three datasets in `.txt` and `.xlsx` formats are contained in this [Zenodo link](https://zenodo.org/records/4632428) and the 1$^o$ x 1$^o$ .


#### * Soued et al.

#### * Hansen et al.
The authors argue that hydropower reservoirs, based on an example from the U.S.A. span a wide range of climate regions and have diverse design and operational characteristics compared to those most heavily represented in model literature (i.e., large, tropical reservoirs). It is therefore not clear whether estimates based on measurements and modeling of other subsets of reservoirs describe those diverse types of hydropower reservoirs. The authors then go on to compare the emission estimates from extrapolations from other measured subsets of reservoirs against direct modelling using G-res. The authors observed that the net GHG reservoir footprint calculated with G-res was less variable and towards the lower end of the range observed from modeling larger global reservoirs, with a range of 138 to 1,052 g CO2 eq m−2 y−1, while the global study reported a range of 115 to 145,472 g CO2 eq m−2 y−1. The authods concluded that due to high variation in emissions normalized with respect to area and generation, we need to be cautious when using area or generation in predicting or communicating emissions footprints for reservoirs relative to those of other energy sources, e.g. in hydropower planning studies. The authors compared their estimates against the published data from [Deemer et al. 2016](https://doi.org/10.1093/biosci/biw117).and the 1$o$ x 1$^o$ gridded estimates of total GHG emissions across all emission pathways [Harrison et al. 2021](https://doi.org/10.1029/2020GB006888). The authors concluded with a good overview of limitations of G-res, particularly with respect to dynamically changing inputs, e.g. due to climate change conditions - see the notes in [Hansen_notes_on_limitations.txt](Hansen_notes_on_limitations.txt)

#### * Tangi et al. 

Tangi et al formulated a robust optimization problem for hydropower planning using GHG emisisons as a novel objective. The authors employed a two-step optimal planning with three objectives : HP generation, sediment transport, and GHG emissions using a multiobjective evolutionary algorithm (MOEA) for obtaining non-dominated (Pareto-optimal) portfolios of dams. The approach is two step: (1) Initial optimization using central estimates of all three objectives producing Pareto-optimal sequences of dams and (2) Portfolio robustness step via screening of solutions to include only robust solutions. 

#### * Carlino et al.



### Notes on limitations of our approach (for the manuscript)

1. We cannot model cascading systems because our approach does not explicitly account for reductions or inputs to GHG emission processes that might be occurring in upstream reservoirs (Prairie et al 2017)

# Package imports
from typing import Tuple, List, Dict, Callable
import pathlib
import os
import numpy as np
import pandas as pd
from geopy.distance import geodesic
import matplotlib.pyplot as plt
import matplotlib.font_manager as font_manager
import matplotlib
import seaborn as sns
sns.set(style='ticks', context='talk')

from sklearn.linear_model import LinearRegression
from sklearn.base import clone
from sklearn.preprocessing import OneHotEncoder
from sklearn.utils import resample
from sklearn.metrics import r2_score

matplotlib.rcParams['pdf.fonttype'] = 3
matplotlib.rcParams['ps.fonttype'] = 42

import warnings
warnings. filterwarnings('ignore', category=UserWarning)

# Dataset imports
harrison_fluxes = pathlib.Path('inputs/gridded_emissions_datasets/harrison2021/G-res_Output_x_Deg_Latitude_20210423.xlsx')
climate_zones_soued = pathlib.Path('inputs/gridded_emissions_datasets/soued2022/1976-2000_ASCII.txt')
soued_emission_factors = pathlib.Path('inputs/gridded_emissions_datasets/soued2022/soued2022.csv')
reemission_file = pathlib.Path('outputs/reemission/outputs_MIN_LOW_PRIM.xlsx')

# Define and create the folder for storing the output figures
fig_folder = pathlib.Path("figures/emission_comparison")
os.makedirs(fig_folder, exist_ok=True)
# Define and create the folder for storing the output files (tabular data)
outfile_folder = pathlib.Path("outputs/emissions_comparison")
os.makedirs(outfile_folder, exist_ok=True)
save_figure_global: bool = True

# Parameters from Eq. 1 - used to disentangle anthropogenic and natural reservoir emissions
R_downstream = 0.17
net_CO2 = 0.25
net_CH4 = 0.90
# Mapping between Koppen-Geiger climate classifications (detailed) and Climatic Zones (coarse)
kopgei_to_climate_cfg = {
    ('Af','Am','As','Aw'): 'Tropical wet moist',
    ('BWh','BSh'): 'Tropical dry montane',
    ('Cfa','Cfb','Cfc','Dfa'): 'Warm temperate moist',
    ('Csa','Csb','Csc','Cwa','Cwb','Cwc'): 'Warm temperate dry',
    ('BWk','BSk','Dfb','Dsa','Dsb','Dwa','Dwb'): 'Cool temperate moist dry',
    ('EF','ET','Dwc','Dwd','Dfc','Dfd','Dsc','Dsd') : 'Polar moist boreal dry moist'
}

# Custom functions

# A function decorator that skips execution of a function. Used to switch execution of functions off without removing code
def skip_execution(reason="Function skipped"):
    def decorator(func):
        def wrapper(*args, **kwargs):
            print(f"{func.__name__} skipped: {reason}")
            # Optionally, you could return None or a custom value if needed
            return None
        return wrapper
    return decorator

def load_climate_zone_soued(filename: str = climate_zones_soued) -> pd.DataFrame:
    """
    The code loads text data with geodesic coordinates and Koppen-Geiger climate classification codes
    Note: Borrowed from Carlino et al. 
    """
    kopgei = pd.read_csv(climate_zones_soued, header=1)
    count = 0
    for col in kopgei.columns[0].split():
        kopgei[col] = [x.split()[count] for x in kopgei[kopgei.columns[0]].values]
        if count < 2:
            kopgei[col] = pd.to_numeric(kopgei[col])
        count += 1
    kopgei = kopgei.drop([kopgei.columns[0]], axis=1)
    return kopgei

def expand_kopgei_climate_mapping(kopgei_climate_map: Dict[Tuple[str, ...], str]) -> Dict[str, str]:
    """ """
    return {item: value for key_tuple, value in kopgei_climate_map.items() for item in key_tuple}

def find_nearest_kopgei_classification(df: pd.DataFrame, lat: float, lon: float, eps: float = 0.25) -> str:
    """Finds the Koppen Geiger classification based on latitude and longitude of a point"""
    # Find rows of data near the location (fast pre-screening)
    df_trimmed = df.loc[
        (abs(df['Lat'].values - lat) <= eps) & \
        (abs(df['Lon'].values - lon) <= eps)]
    # Calculate geodesic distance for each point in the dataframe (slow on large dataframes)
    distances = df_trimmed.apply(lambda row: geodesic((lat, lon), (row['Lat'], row['Lon'])).kilometers, axis=1)
    # Find the index of the minimum distance
    nearest_index = distances.idxmin()
    # Return the classification of the closest point
    return df.loc[nearest_index, 'Cls']

def add_emission_factors(em_factors: pd.DataFrame, koppen_geiger: str): 
    czone = kopgei_to_czone[koppen_geiger]
    data_dict = em_factors[em_factors['Climate']==czone].iloc[0].to_dict()
    return data_dict

# Error functions
def _abs_err(x_1, x_2) -> float | np.float64:
    return np.abs(x_1-x_2)

def _rel_err(x_1, x_2) -> float | np.float64:
    return (x_2 - x_1) / x_1

def _rel_err2(y_true, y_pred) -> float | np.float64:
    return (y_true-y_pred) / y_true

def _rel_abs_err(x_1, x_2) -> float | np.float64:
    return _abs_err(x_1, x_2) / x_1

# x1 is true value
# x2 is prediction

# Mean Absolute Error (MAE)
def mae(x_1, x_2) -> float | np.float64:
    return _abs_err(x_1, x_2).mean()

# Root Mean Squared Error (RMSE)
def rmse(x_1, x_2) -> float | np.float64:
    return np.sqrt((_abs_err(x_1, x_2)**2).mean())

# Mean Absolute Percentage Error
def mape(x_1, x_2) -> float | np.float64:
    return _rel_abs_err(x_1, x_2).mean() * 100

# Relative Mean Absolute Error
def rmae(x_1, x_2) -> float | np.float64:
    return _rel_abs_err(x_1, x_2).mean()
    
# Relative Root Mean Squared Error (RRMSE)
def rrmse(x_1, x_2) -> float | np.float64:
    return np.sqrt((_rel_abs_err(x_1, x_2) ** 2).mean())

# Symmetric Mean Absolute Percentage Error (SMAPE)
def smape(x_1, x_2) -> float | np.float64:
    return (_abs_err(x_1, x_2) / ((np.abs(x_1) + np.abs(x_2)) / 2)).mean() * 100

def r_squared(x_1, x_2) -> np.float64:
    """x_1 = y_true, x_2 = y_pred"""
    # Convert inputs to numpy arrays for calculation
    x_1 = np.array(x_1)
    x_2 = np.array(x_2)
    # Calculate mean of full model outputs
    x_1_mean = np.mean(x_1)
    # Calculate SSE (Sum of Squared Errors) and SST (Total Sum of Squares)
    sse = np.sum((x_1 - x_2) ** 2)
    sst = np.sum((x_1 -x_1_mean) ** 2)
    # Calculate R-squared
    r2 = 1 - (sse / sst)
    return r2

def report_errors(x_1, x_2) -> None:
    print(f"Mean Absolute Error (MAE): {mae(x_1, x_2)}")
    print(f"Coefficient of determination (R-squared): {r_squared(x_1, x_2)}")
    print(f"Mean Absolute Percentage Error (MAPE): {mape(x_1, x_2)}")
    print(f"Root Mean Squared Error (RMSE): {rmse(x_1, x_2)}")
    print(f"Relative Mean Absolute Error (RMAE): {rmae(x_1, x_2)}")
    print(f"Relative Root Mean Squared Error (RRMSE): {rrmse(x_1, x_2)}")
    print(f"Symmetric Mean Absolute Percentage Error (SMAPE): {smape(x_1, x_2)}")
    
# Compare total emissions
def total_emission(row, em_column: str, area_column: str, conv_coefficient: float = 1.0) -> None:
    """Calculate total emissions. Given that emission intensities are in gCO2e/m2/year and areas in km2, the
    calculated total emission at `conv_coefficient = 1.0` is in the unit of tonnesCO2e/year."""
    return row[em_column] * row[area_column] * conv_coefficient

## Preprocess the gridded emission datasets

### * Parameterization on the basis of latitude by Harrison et al. (2021)
- **NOTE:** We are not using those as they the data is for a single point in time (year 2020) whilst we are looking for average emissions over the lifetime of a reservoir (e.g. 100 years)

# Read GHG fluxes from Harrison et al. (2021)
GHGFluxes = (pd.read_excel(harrison_fluxes).
                assign(
                    CO2_net=lambda x: x['avg.CO2.diff.g.m2.y.CO2.eq.x.lat'] * net_CO2 * (1 + R_downstream),
                    CH4_net=lambda x: x['avg.CH4.tot.g.m2.y.CO2.eq.x.lat'] * net_CH4 * (1 + R_downstream),
                    tot_em_net=lambda x: x['CO2_net'] + x['CH4_net']).
                 drop([
                     'sum.CH4.tot.Mg.CO2.eq.x.lat', 'sum.CH4.deg.Mg.CO2.eq.x.lat', 
                     'sum.CH4.eb.Mg.CO2.eq.x.lat', 'sum.CH4.diff.Mg.CO2.eq.x.lat', 
                     'sum.CO2.diff.Mg.CO2.eq.x.lat', 'sum.C.gas.tot.Mg.CO2.eq.x.lat'],
                     axis=1)
            )
GHGFluxes.head()

### * Parameterization on the basis of climatic zones by Soued et al. (2022)
* **NOTE**: Related to the refined emission factors from the 2019 IPCC Guidelines

# Calculate areal emissions from climate zone parameterization included in Soued et al. 2022; 
# The code was adapted from Carlino et al.
# Read Koppen-Geiger climate zone classification codes per location
kopgei = load_climate_zone_soued(filename = climate_zones_soued)
# Get a list of all Koppen-Geiger classifications
kopgeizone = kopgei['Cls'].unique().tolist()
# Load emission factors per climatic zones
soued2022 = pd.read_csv(soued_emission_factors).dropna()
czone = soued2022['Climate'].tolist()
# Create a mapping betwee Koppen-Geiger codes and coarser climate zone classifications used for 
kopgei_to_czone = expand_kopgei_climate_mapping(kopgei_to_climate_cfg)
soued2022.head(10)

## Load and process emission outputs calculated with ReEmission

* **NOTE:** `res_area` and `catch_area` are given in km$^2$

input_columns_to_keep = [
    'Name', 'coordinates_0', 'coordinates_1', 'id', 'type', 'catch_area', 
    'res_area', 'res_volume']
reemission_inputs = pd.read_excel(reemission_file, sheet_name="inputs", usecols=input_columns_to_keep)
output_columns_to_keep = ['Name', 'co2_diffusion', 'ch4_diffusion', 'ch4_ebullition', 'ch4_degassing', 'co2_net', 'ch4_net']
reemission_outputs = pd.read_excel(reemission_file, sheet_name="outputs", usecols=output_columns_to_keep)
# Join both dataframes on Name
reemission_df = pd.merge(reemission_inputs, reemission_outputs, on='Name', how='inner')
# Double check that we have not lost any rows
assert len(reemission_df) == len(reemission_inputs) == len(reemission_outputs)
# Add climate classification value for each row
reemission_df['Cls'] = reemission_df.apply(
    lambda row: find_nearest_kopgei_classification(
        df = kopgei, lat=row['coordinates_0'], lon=row['coordinates_1']), axis=1)
# Apply the function to create a DataFrame from the list returned by add_blank
soued_em = reemission_df.apply(
    lambda row: add_emission_factors(soued2022, row['Cls']), axis='columns', result_type='expand')
reemission_df = pd.concat([reemission_df, soued_em], axis=1).\
    rename(columns={
        'CO2 diffusive 20 yrs integrated [gCO2eq/ (m2 yr)]': 'co2_diffusion_soued',
        'CH4 diffusive 20 yrs integrated [gCO2eq/ (m2 yr)]': 'ch4_diffusion_soued',
        'CH4 ebullition [gCO2eq/ (m2 yr)]': 'ch4_ebullition_soued',
        'CH4 degassing [gCO2eq/ (m2 yr)]': 'ch4_degassing_soued'})
reemission_df.describe().drop('count')

## Perform a linear regression to find the values of emission factors with the best fit to data (outputs from ReEmission)

# Find out if the data is balanced with respect to the climate categorical variables
climate_counts = reemission_df['Climate'].value_counts().to_dict()
print(climate_counts)
# The data looks imbalanced with respect to climate categories and therefore we shall employ
# a weighted linear regressions with weights inversely proportional to the value counts for each
# category
# Prepare the data for solving the linear regression problem
coeff = 1e-3 # Unit conversion
regression_df = (reemission_df.copy().
    assign(co2_diffusion_total=lambda x: total_emission(x, 'co2_diffusion', 'res_area', coeff)).
    assign(ch4_diffusion_total=lambda x: total_emission(x, 'ch4_diffusion', 'res_area', coeff)).
    assign(ch4_ebullition_total=lambda x: total_emission(x, 'ch4_ebullition', 'res_area', coeff)).
    assign(ch4_degassing_total=lambda x: total_emission(x, 'ch4_degassing', 'res_area', coeff))
)

def get_regression_data(
        df: pd.DataFrame, 
        output_column: str, 
        area_col: str = 'res_area') -> Tuple[pd.DataFrame, pd.Series, pd.Series]:
    """Returns input-output data for fitting a linear regression model"""
    encoder = OneHotEncoder()  # Drop one category to avoid multicollinearity
    climate_encoded = encoder.fit_transform(df[['Climate']]).toarray()
    encoded_climate_df = pd.DataFrame(
        climate_encoded, 
        columns=encoder.get_feature_names_out(['Climate']))
    X = encoded_climate_df
    #X = pd.concat([encoded_climate_df, df[area_col]], axis=1)
    y = df[output_column] / df[area_col]
    # Calculate sample weights based on inverse frequency of each category
    climate_counts = df['Climate'].value_counts()
    weights = df['Climate'].map(lambda x: 1 / climate_counts[x])
    return X, y, weights
    
def find_emission_factors(
        X: pd.DataFrame, 
        y: pd.Series, 
        weights: pd.Series, 
        coeff: float = 1e-3):
    """Finds emission coefficients using weighted linear regression between climate zone and emission
    predictions calculated with G-res.
    """
    # Initialize the model
    model = LinearRegression(fit_intercept=False)
    # Fit the model on the original data to get the initial coefficients
    model.fit(X, y, sample_weight=weights)
    base_coefficients = model.coef_ / coeff
    coeff_dict = dict(zip(X.columns, base_coefficients))
    return model, coeff_dict

def find_confidence_intervals(
        X: pd.DataFrame,
        y: pd.Series,
        model,
        weights: pd.Series = None,
        coeff: float = 1e-3,
        n_bootstraps: int = 1_000, 
        confidence_level: float = 0.95) -> pd.Series:
    """ """
    # Store bootstrap coefficients
    bootstrap_coefficients = []
    # Perform bootstrapping
    for i in range(n_bootstraps):
        model_cf = clone(model)
        # Resample data with replacement
        if weights:
            X_resampled, y_resampled, weights_resampled = resample(X, y, weights)
            # Fit model on the resampled data
            model_cf.fit(X_resampled, y_resampled, sample_weight=weights_resampled)
        else:
            X_resampled, y_resampled = resample(X, y)
            model_cf.fit(X_resampled, y_resampled)
        base_coefficients = model_cf.coef_ / coeff
        # Store the coefficients
        bootstrap_coefficients.append(base_coefficients)
        
    # Convert bootstrap coefficients to a DataFrame
    bootstrap_coefficients = np.array(bootstrap_coefficients)
    bootstrap_df = pd.DataFrame(bootstrap_coefficients, columns=X.columns)
    # Calculate the confidence intervals for each coefficient
    ci_lower = bootstrap_df.quantile((1 - confidence_level) / 2)
    ci_upper = bootstrap_df.quantile(1 - (1 - confidence_level) / 2)
    deltas = (ci_upper - ci_lower) / 2
    return deltas

## Run regressions for all emission pathways

# Run regressions for all emission pathways
use_weights: bool = False
verbose: bool = False
exec_options = {
    "CO2 diffusion": "co2_diffusion_total",
    "CH4 diffusion": "ch4_diffusion_total",
    "CH4 ebullition": "ch4_ebullition_total",
    "CH4 degassing": "ch4_degassing_total"
}
coefficients_dict = {}
ci_deltas_dict = {}
r2_scores_dict = {}
for pathway_name, pathway_column in exec_options.items():
    X, y, weights = get_regression_data(regression_df, pathway_column)
    if not use_weights:
        weights = pd.Series(1, index=regression_df.index)
        weights /= len(weights)
    model, coeffs = find_emission_factors(X, y, weights)
    if verbose:
        print(coeffs)
    coefficients_dict[pathway_name] = coeffs
    if use_weights:
        deltas = find_confidence_intervals(X, y, model, weights)
        r2 = r2_score(y, model.predict(X), sample_weight=weights)
    else:
        deltas = find_confidence_intervals(X, y, model)
        r2 = r2_score(y, model.predict(X))
    r2_scores_dict[pathway_name] = r2
    ci_deltas_dict[pathway_name] = deltas
    
coeffs_df_fitted = pd.DataFrame(coefficients_dict)\
    .reset_index().rename(columns={"index": "Climate"})\
    .rename(columns = {
        "CO2 diffusion": "co2_diffusion_fitted",
        "CH4 diffusion": "ch4_diffusion_fitted",
        "CH4 ebullition": "ch4_ebullition_fitted",
        "CH4 degassing": "ch4_degassing_fitted"
    })
coeffs_df_fitted["Climate"] = coeffs_df_fitted["Climate"].apply(lambda x: x.replace("Climate_", ""))
coeffs_df_fitted

pd.DataFrame(r2_scores_dict, index=["R2 fitted"])

pd.DataFrame(ci_deltas_dict)

### Add columns with the fitted emission factors (EFs) to the `reemission_df` dataframe

fitted_em = reemission_df.apply(
    lambda row: add_emission_factors(coeffs_df_fitted, row['Cls']), axis='columns', result_type='expand')
reemission_df = pd.concat([reemission_df, fitted_em], axis=1)

reemission_df.head()

# Compare emission intensities
report = (reemission_df.
              describe().
              loc[
                  :, 
                  ['co2_diffusion', 'co2_diffusion_soued', 'co2_diffusion_fitted',
                   'ch4_diffusion', 'ch4_diffusion_soued', 'ch4_diffusion_fitted',
                   'ch4_ebullition', 'ch4_ebullition_soued', 'ch4_ebullition_fitted', 
                   'ch4_degassing', 'ch4_degassing_soued', 'ch4_degassing_fitted']
              ].
              T.drop_duplicates().T.
              drop('count')
         )
report

def remove_duplicated_columns(df: pd.DataFrame) -> pd.DataFrame:
    return df.loc[:,~df.columns.duplicated()].copy()
reemission_df = remove_duplicated_columns(reemission_df)

### Calculate total emissions from Emission Factors (EFs) - both Soued's and the fitted ones

coeff = 1e-3
emissions_total = (reemission_df.
    assign(co2=lambda x: x['co2_diffusion']).
    assign(co2_soued=lambda x: x['co2_diffusion_soued']).
    assign(co2_fitted=lambda x: x['co2_diffusion_fitted']).               
    assign(ch4=lambda x: x['ch4_diffusion'] + x['ch4_ebullition'] + x['ch4_degassing']).
    assign(ch4_soued=lambda x: x['ch4_diffusion_soued'] + x['ch4_ebullition_soued'] + x['ch4_degassing_soued']).
    assign(ch4_fitted=lambda x: x['ch4_diffusion_fitted'] + x['ch4_ebullition_fitted'] + x['ch4_degassing_fitted']).               
    assign(co2_net_total=lambda x: total_emission(x, 'co2_net', 'res_area', coeff)).
    assign(ch4_net_total=lambda x: total_emission(x, 'ch4_net', 'res_area', coeff)).
    assign(co2_diffusion_total=lambda x: total_emission(x, 'co2_diffusion', 'res_area', coeff)).
    assign(co2_diffusion_total_soued=lambda x: total_emission(x, 'co2_diffusion_soued', 'res_area', coeff)).
    assign(co2_diffusion_total_fitted=lambda x: total_emission(x, 'co2_diffusion_fitted', 'res_area', coeff)).               
    assign(ch4_diffusion_total=lambda x: total_emission(x, 'ch4_diffusion', 'res_area', coeff)).
    assign(ch4_diffusion_total_soued=lambda x: total_emission(x, 'ch4_diffusion_soued', 'res_area', coeff)).
    assign(ch4_diffusion_total_fitted=lambda x: total_emission(x, 'ch4_diffusion_fitted', 'res_area', coeff)).
    assign(ch4_ebullition_total=lambda x: total_emission(x, 'ch4_ebullition', 'res_area', coeff)).
    assign(ch4_ebullition_total_soued=lambda x: total_emission(x, 'ch4_ebullition_soued', 'res_area', coeff)).
    assign(ch4_ebullition_total_fitted=lambda x: total_emission(x, 'ch4_ebullition_fitted', 'res_area', coeff)).               
    assign(ch4_degassing_total=lambda x: total_emission(x, 'ch4_degassing', 'res_area', coeff)).
    assign(ch4_degassing_total_soued=lambda x: total_emission(x, 'ch4_degassing_soued', 'res_area', coeff)).
    assign(ch4_degassing_total_fitted=lambda x: total_emission(x, 'ch4_degassing_fitted', 'res_area', coeff)).               
    assign(co2_total = lambda x: x['co2_diffusion_total']).
    assign(co2_total_soued = lambda x: x['co2_diffusion_total_soued']).
    assign(co2_total_fitted = lambda x: x['co2_diffusion_total_fitted']).               
    assign(ch4_total = lambda x : x['ch4_diffusion_total'] + x['ch4_ebullition_total'] + x['ch4_degassing_total']).
    assign(ch4_total_soued = lambda x : x['ch4_diffusion_total_soued'] + x['ch4_ebullition_total_soued'] + x['ch4_degassing_total_soued']).
    assign(ch4_total_fitted = lambda x : x['ch4_diffusion_total_fitted'] + x['ch4_ebullition_total_fitted'] + x['ch4_degassing_total_fitted']).            
    #T.drop([
    #    'co2_diffusion', 'co2_net', 'ch4_diffusion', 'ch4_net', 'ch4_ebullition', 'ch4_degassing',
    #    'co2_diffusion_soued', 'ch4_diffusion_soued', 'ch4_ebullition_soued', 'ch4_degassing_soued']).T.
    loc[:, [
            'co2_diffusion', 'co2_net', 'ch4_diffusion', 'ch4_net', 'ch4_ebullition', 'ch4_degassing',
            'co2_diffusion_soued', 'ch4_diffusion_soued', 'ch4_ebullition_soued', 'ch4_degassing_soued',
            'co2_diffusion_fitted', 'ch4_diffusion_fitted', 'ch4_ebullition_fitted', 'ch4_degassing_fitted',
            'co2', 'co2_soued', 'co2_fitted', 'ch4', 'ch4_soued', 'ch4_fitted',
            'co2_diffusion_total', 'co2_diffusion_total_soued', 'co2_diffusion_total_fitted',
            'ch4_diffusion_total', 'ch4_diffusion_total_soued', 'ch4_diffusion_total_fitted',
            'ch4_ebullition_total', 'ch4_ebullition_total_soued', 'ch4_ebullition_total_fitted',
            'ch4_degassing_total', 'ch4_degassing_total_soued', 'ch4_degassing_total_fitted',
            'co2_total', 'co2_total_soued', 'co2_total_fitted',
            'ch4_total', 'ch4_total_soued', 'ch4_total_fitted', 
            'co2_net_total', 'ch4_net_total',
            'type', 'res_area', 'res_volume', 'Climate']]
)
emissions_total = emissions_total.astype(
    {col: 'float' for col in emissions_total.columns if col not in ('type', 'Climate')})
emissions_total['type'] = emissions_total['type'].astype('category')
emissions_total['Climate'] = emissions_total['Climate'].astype('category')
emissions_total = emissions_total.loc[:, ~emissions_total.columns.duplicated()]
emissions_total.describe().drop('count')

### Calculate basic statistics

pd.DataFrame(emissions_total[[
            'co2_diffusion_total', 'co2_diffusion_total_soued', 'co2_diffusion_total_fitted', 
            'ch4_diffusion_total', 'ch4_diffusion_total_soued', 'ch4_diffusion_total_fitted', 
            'ch4_ebullition_total', 'ch4_ebullition_total_soued', 'ch4_ebullition_total_fitted',
            'ch4_degassing_total', 'ch4_degassing_total_soued', 'ch4_degassing_total_fitted',
            'co2_total', 'co2_total_soued', 'co2_total_fitted',
            'ch4_total', 'ch4_total_soued', 'ch4_total_fitted', 
            'co2_net_total', 'ch4_net_total']].sum()).T

pd.DataFrame(emissions_total[[
    'co2_diffusion', 'co2_diffusion_soued', 'co2_diffusion_fitted', 
    'ch4_diffusion', 'ch4_diffusion_soued', 'ch4_diffusion_fitted',
    'ch4_ebullition', 'ch4_ebullition_soued', 'ch4_ebullition_fitted',
    'ch4_degassing', 'ch4_degassing_soued', 'ch4_degassing_fitted']].mean()).T

pd.DataFrame(emissions_total[[
    'co2_diffusion', 'co2_diffusion_soued', 'co2_diffusion_fitted', 
    'ch4_diffusion', 'ch4_diffusion_soued', 'ch4_diffusion_fitted',
    'ch4_ebullition', 'ch4_ebullition_soued', 'ch4_ebullition_fitted',
    'ch4_degassing', 'ch4_degassing_soued', 'ch4_degassing_fitted', 'Climate']].groupby("Climate").mean())

## Make comparison plots

pairs = [
    ('co2_diffusion_total', 'co2_diffusion_total_soued'), 
    ('ch4_diffusion_total', 'ch4_diffusion_total_soued'), 
    ('co2_total', 'co2_total_soued'), 
    ('ch4_ebullition_total', 'ch4_ebullition_total_soued'), 
    ('ch4_degassing_total', 'ch4_degassing_total_soued'), 
    ('ch4_total', 'ch4_total_soued')
]
axis_labels = {
    'co2_diffusion_total' : 'CO$_2$ diffusion (model), ktCO$_{2e}$/yr',
    'co2_diffusion_total_soued': 'CO$_2$ diffusion (EF), ktCO$_{2e}$/yr',
    'ch4_diffusion_total': 'CH$_4$ diffusion (model), ktCO$_{2e}$/yr',
    'ch4_diffusion_total_soued': 'CH$_4$ diffusion (EF), ktCO$_{2e}$/yr',
    'ch4_ebullition_total': 'CH$_4$ ebullition (model), ktCO$_{2e}$/yr',
    'ch4_ebullition_total_soued': 'CH$_4$ ebullition (EF), ktCO$_{2e}$/yr',
    'ch4_degassing_total': 'CH$_4$ degassing (model), ktCO$_{2e}$/yr',
    'ch4_degassing_total_soued': 'CH$_4$ degassing (EF), ktCO$_{2e}$/yr',
    'co2_total': 'CO$_2$ emission (model), ktCO$_{2e}$/yr',
    'co2_total_soued': 'CO$_2$ emission (EF), ktCO$_{2e}$/yr',
    'ch4_total': 'CH$_4$ emission (model), ktCO$_{2e}$/yr',
    'ch4_total_soued': 'CH$_4$ emission (EF), ktCO$_{2e}$/yr'}
suptitle = "Comparison of Emissions: Explicit Calculations vs. Emission Factor-Based Estimates"
exec_message = "This plot is replaced with a 4 x 2 grid plot with top row using established EFs " + \
    "and EFs calculated via regression using G-res calculated emissions"
@skip_execution(exec_message)
def make_pair_plots_total_emissions(
        df: pd.DataFrame, 
        pairs, axis_labels, suptitle,
        filename: str,
        dxdy1: float = 0.0, dxdy2: float = 300, 
        save_fig: bool = save_figure_global) -> None:
    # Define the pairs of columns to plot
    sns.set_theme(style="ticks")
    sns.set_context("paper", font_scale=1.6) 
    label_fontsize = 15
    tick_fontsize = 15

    # Create the grid for subplots
    fig, axes = plt.subplots(nrows=2, ncols=3, figsize=(15, 10))
    axes = axes.flatten()
    # Generate q-q plots for each pair
    for idx, (col_x, col_y) in enumerate(pairs):
        ax = axes[idx]
        legend_status='auto'
        sns.scatterplot(
            x=col_x, y=col_y, style = 'type', hue='type', size='res_area', ax=ax, data=df,
            markers=['o', 's', '^'], alpha=0.4, sizes=(50, 500),
            **{'edgecolor':'black'})
        ax.set_xscale('log')
        ax.set_yscale('log')
        max_val = max(df[[col_x, col_y]].max())
        min_val = min(df[[col_x, col_y]].min())
        ax.plot([min_val, max_val], [min_val, max_val], 'k--', lw=3, alpha = 0.6)
        # Customize each subplot
        ax.set_xlabel(axis_labels[col_x])
        ax.set_ylabel(axis_labels[col_y])
        #ax.set_xticklabels(ax.get_xticks(), size = tick_fontsize)
        #ax.set_yticklabels(ax.get_yticks(), size = tick_fontsize)
        ax.set_ylim([min_val-dxdy1, max_val+dxdy2])
        ax.set_xlim([min_val-dxdy1, max_val+dxdy2])

    handles, labels = ax.get_legend_handles_labels()
    for handle in handles:
        if handle._label in ('hydroelectric', 'multipurpose', 'irrigation'):
            pass
            #handle.set_markersize(handle.get_markersize() * 2)
            
    fig.legend(
        handles, [label.replace('type', 'Reservoir Type').replace('res_area', 'Area, km$^2$') for label in labels],
        loc='right', bbox_to_anchor=(1.085, 0.5), 
        scatterpoints=1, frameon=False)

    # Iterate again and remove legend from each subplot
    for idx, _ in enumerate(pairs):
        ax = axes[idx]
        ax.legend().remove()
    plt.tight_layout(rect=[0, 0, 0.95, 0.95])  # Adjust layout to make space for the legend
    fig.suptitle(
        suptitle, 
        fontsize=18, fontweight='bold')
    if save_fig:
        fig.savefig(fig_folder / f'{filename}.pdf', bbox_inches='tight')
        fig.savefig(fig_folder / f'{filename}.svg', bbox_inches='tight')
    plt.show()
make_pair_plots_total_emissions(
    emissions_total, 
    pairs, 
    axis_labels, 
    suptitle, 
    filename="emission_pathways_comparison")

pairs_fitted = [
    ('co2_diffusion_total', 'co2_diffusion_total_fitted'), 
    ('ch4_diffusion_total', 'ch4_diffusion_total_fitted'), 
    ('co2_total', 'co2_total_fitted'), 
    ('ch4_ebullition_total', 'ch4_ebullition_total_fitted'), 
    ('ch4_degassing_total', 'ch4_degassing_total_fitted'), 
    ('ch4_total', 'ch4_total_fitted')
]
axis_labels_fitted = {
    'co2_diffusion_total' : 'CO$_2$ diffusion (model), ktCO$_{2e}$/yr',
    'co2_diffusion_total_fitted': 'CO$_2$ diffusion (EF), ktCO$_{2e}$/yr',
    'ch4_diffusion_total': 'CH$_4$ diffusion (model), ktCO$_{2e}$/yr',
    'ch4_diffusion_total_fitted': 'CH$_4$ diffusion (EF), ktCO$_{2e}$/yr',
    'ch4_ebullition_total': 'CH$_4$ ebullition (model), ktCO$_{2e}$/yr',
    'ch4_ebullition_total_fitted': 'CH$_4$ ebullition (EF), ktCO$_{2e}$/yr',
    'ch4_degassing_total': 'CH$_4$ degassing (model), ktCO$_{2e}$/yr',
    'ch4_degassing_total_fitted': 'CH$_4$ degassing (EF), ktCO$_{2e}$/yr',
    'co2_total': 'CO$_2$ emission (model), ktCO$_{2e}$/yr',
    'co2_total_fitted': 'CO$_2$ emission (EF), ktCO$_{2e}$/yr',
    'ch4_total': 'CH$_4$ emission (model), ktCO$_{2e}$/yr',
    'ch4_total_fitted': 'CH$_4$ emission (EF), ktCO$_{2e}$/yr'}
suptitle_fitted = "Comparison of Emissions: Explicit Calculations vs. Emission Factor-Based Estimates with fitted coefficients"
make_pair_plots_total_emissions(
    emissions_total, 
    pairs_fitted, 
    axis_labels_fitted, 
    suptitle_fitted, 
    filename="emission_pathways_comparison_fitted")

print("Columns in emissions_total dataframe")
print("------------------------------------")
print("\n".join(emissions_total.columns))
print("------------------------------------")

@skip_execution("These plots are not used in the publication.")
def make_pair_plots_emission_fluxes(
        df: pd.DataFrame, dxdy1: float = 0.0, dxdy2: float = 300,
        save_fig: bool = save_figure_global) -> None:
    # Define the pairs of columns to plot
    sns.set_theme(style="ticks")
    sns.set_context("paper", font_scale=1.6) 
    label_fontsize = 15
    tick_fontsize = 15
    pairs = [
        ('co2_diffusion', 'co2_diffusion_soued'), 
        ('ch4_diffusion', 'ch4_diffusion_soued'), 
        ('co2', 'co2_soued'), 
        
        ('ch4_ebullition', 'ch4_ebullition_soued'), 
        ('ch4_degassing', 'ch4_degassing_soued'), 
        ('ch4', 'ch4_soued')
    ]
    axis_labels = {
        'co2_diffusion' : 'CO$_2$ diffusion flux (model), gCO$_{2e}$m$^2$/yr',
        'co2_diffusion_soued': 'CO$_2$ diffusion flux (EF), gCO$_{2e}$m$^2$/yr',
        'ch4_diffusion': 'CH$_4$ diffusio fluxn (model), gCO$_{2e}$m$^2$/yr',
        'ch4_diffusion_soued': 'CH$_4$ diffusion flux (EF), gCO$_{2e}$m$^2$/yr',
        'ch4_ebullition': 'CH$_4$ ebullition flux (model), gCO$_{2e}$m$^2$/yr',
        'ch4_ebullition_soued': 'CH$_4$ ebullition flux (EF), gCO$_{2e}$m$^2$/yr',
        'ch4_degassing': 'CH$_4$ degassing flux (model), gCO$_{2e}$m$^2$/yr',
        'ch4_degassing_soued': 'CH$_4$ degassing flux (EF), gCO$_{2e}$m$^2$/yr',
        'co2': 'CO$_2$ emission flux (model), gCO$_{2e}$m$^2$/yr',
        'co2_soued': 'CO$_2$ emission flux (EF), gCO$_{2e}$m$^2$/yr',
        'ch4': 'CH$_4$ emission flux (model), gCO$_{2e}$m$^2$/yr',
        'ch4_soued': 'CH$_4$ emission flux (EF), gCO$_{2e}$m$^2$/yr'}
    # Create the grid for subplots
    fig, axes = plt.subplots(nrows=2, ncols=3, figsize=(15, 10))
    axes = axes.flatten()
    # Generate q-q plots for each pair
    for idx, (col_x, col_y) in enumerate(pairs):
        ax = axes[idx]
        legend_status='auto'
        sns.scatterplot(
            x=col_x, y=col_y, style = 'type', hue='type', size='res_area', ax=ax, data=df,
            markers=['o', 's', '^'], alpha=0.5, sizes=(50, 500),
            **{'edgecolor':'black'})
        #ax.set_xscale('log')
        #ax.set_yscale('log')
        max_val = max(df[[col_x, col_y]].max())
        min_val = min(df[[col_x, col_y]].min())
        ax.plot([min_val, max_val], [min_val, max_val], 'k--', lw=3, alpha = 0.6)
        # Customize each subplot
        ax.set_xlabel(axis_labels[col_x])
        ax.set_ylabel(axis_labels[col_y])
        #ax.set_xticklabels(ax.get_xticks(), size = tick_fontsize)
        #ax.set_yticklabels(ax.get_yticks(), size = tick_fontsize)
        ax.set_ylim([min_val-dxdy1, max_val+dxdy2])
        ax.set_xlim([min_val-dxdy1, max_val+dxdy2])

    handles, labels = ax.get_legend_handles_labels()
    for handle in handles:
        if handle._label in ('hydroelectric', 'multipurpose', 'irrigation'):
            handle.set_markersize(handle.get_markersize() * 2)
            
    fig.legend(
        handles, [label.replace('type', 'Type').replace('res_area', 'Area, km$^2$') for label in labels],
        loc='right', bbox_to_anchor=(1.085, 0.48), 
        scatterpoints=1, frameon=False)

    # Iterate again and remove legend from each subplot
    for idx, _ in enumerate(pairs):
        ax = axes[idx]
        ax.legend().remove()
    
    plt.tight_layout(rect=[0, 0, 0.95, 0.85])  # Adjust layout to make space for the legend
    if save_fig:
        fig.savefig(fig_folder / 'emission_flux_pathways_comparison.svg', bbox_inches='tight')
        fig.savefig(fig_folder / 'emission_flux_pathways_comparison.pdf', bbox_inches='tight')
    plt.show()
make_pair_plots_emission_fluxes(df=emissions_total)

### Make the publication figure and save to `png`, `svg` and `pdf` files

# Make EF vs G-res comparison figure for the paper
pairs_publication = [
    ('co2_diffusion_total', 'co2_diffusion_total_soued'), 
    ('ch4_diffusion_total', 'ch4_diffusion_total_soued'), 
    ('ch4_ebullition_total', 'ch4_ebullition_total_soued'), 
    ('ch4_degassing_total', 'ch4_degassing_total_soued'), 
    ('co2_diffusion_total', 'co2_diffusion_total_fitted'), 
    ('ch4_diffusion_total', 'ch4_diffusion_total_fitted'),  
    ('ch4_ebullition_total', 'ch4_ebullition_total_fitted'), 
    ('ch4_degassing_total', 'ch4_degassing_total_fitted')
]
axis_labels_publication = {
    'co2_diffusion_total' : 'CO$_2$ diffusion (model), ktCO$_{2e}$/yr',
    'co2_diffusion_total_fitted': 'CO$_2$ diffusion (EF - fitted), ktCO$_{2e}$/yr',
    'ch4_diffusion_total': 'CH$_4$ diffusion (model), ktCO$_{2e}$/yr',
    'ch4_diffusion_total_fitted': 'CH$_4$ diffusion (EF - fitted), ktCO$_{2e}$/yr',
    'ch4_ebullition_total': 'CH$_4$ ebullition (model), ktCO$_{2e}$/yr',
    'ch4_ebullition_total_fitted': 'CH$_4$ ebullition (EF - fitted), ktCO$_{2e}$/yr',
    'ch4_degassing_total': 'CH$_4$ degassing (model), ktCO$_{2e}$/yr',
    'ch4_degassing_total_fitted': 'CH$_4$ degassing (EF - fitted), ktCO$_{2e}$/yr',
    'co2_diffusion_total_soued': 'CO$_2$ diffusion (EF), ktCO$_{2e}$/yr',
    'ch4_diffusion_total_soued': 'CH$_4$ diffusion (EF), ktCO$_{2e}$/yr',
    'ch4_ebullition_total_soued': 'CH$_4$ ebullition (EF), ktCO$_{2e}$/yr',
    'ch4_degassing_total_soued': 'CH$_4$ degassing (EF), ktCO$_{2e}$/yr'}
suptitle_publication = "Comparison of Total Emissions in ktCO$_{2e}$/yr: Explicit Calculations vs. Emission Factor-Based Estimates"

def make_pair_plots_total_emissions_publication(
        df: pd.DataFrame, 
        pairs, axis_labels, suptitle,
        filename: str,
        dxdy1: float = 0.0, dxdy2: float = 300, 
        save_fig: bool = save_figure_global) -> None:
    # Define the pairs of columns to plot
    sns.set_theme(style="ticks")
    sns.set_context("paper", font_scale=1.6) 
    label_fontsize = 12
    tick_fontsize = None

    # Create the grid for subplots
    fig, axes = plt.subplots(nrows=2, ncols=4, figsize=(15, 7))
    axes = axes.flatten()
    # Generate q-q plots for each pair
    for idx, (col_x, col_y) in enumerate(pairs):
        ax = axes[idx]
        legend_status='auto'
        sns.scatterplot(
            x=col_x, y=col_y, style = 'type', hue='type', size='res_area', ax=ax, data=df,
            markers=['o', 's', '^'], alpha=0.4, sizes=(50, 500),
            **{'edgecolor':'black'})
        ax.set_xscale('log')
        ax.set_yscale('log')
        max_val = max(df[[col_x, col_y]].max())
        min_val = min(df[[col_x, col_y]].min())
        ax.plot([min_val, max_val], [min_val, max_val], 'k--', lw=3, alpha = 0.6)
        # Customize each subplot
        ax.set_xlabel(axis_labels[col_x], fontsize=label_fontsize)
        ax.set_ylabel(axis_labels[col_y], fontsize = label_fontsize)
        if tick_fontsize:
            ax.set_xticklabels(ax.get_xticks(), size = tick_fontsize)
            ax.set_yticklabels(ax.get_yticks(), size = tick_fontsize)
        ax.set_ylim([min_val-dxdy1, max_val+dxdy2])
        ax.set_xlim([min_val-dxdy1, max_val+dxdy2])

    handles, labels = ax.get_legend_handles_labels()
    for handle in handles:
        if handle._label in ('hydroelectric', 'multipurpose', 'irrigation'):
            pass
            #handle.set_markersize(handle.get_markersize() * 2)
            
    fig.legend(
        handles, [label.replace('type', 'Reservoir Type').replace('res_area', 'Area, km$^2$') for label in labels],
        loc='right', bbox_to_anchor=(1.085, 0.5), 
        scatterpoints=1, frameon=False)

    # Iterate again and remove legend from each subplot
    for idx, _ in enumerate(pairs):
        ax = axes[idx]
        ax.legend().remove()
    plt.tight_layout(rect=[0, 0, 0.95, 0.95])  # Adjust layout to make space for the legend
    fig.suptitle(
        suptitle, 
        fontsize=18, fontweight='bold')
    if save_fig:
        fig.savefig(fig_folder / f'{filename}.pdf', bbox_inches='tight')
        fig.savefig(fig_folder / f'{filename}.svg', bbox_inches='tight')
        fig.savefig(fig_folder / f'{filename}.png', bbox_inches='tight', dpi=1200)
    plt.show()
make_pair_plots_total_emissions_publication(
    emissions_total, 
    pairs_publication, 
    axis_labels_publication, 
    suptitle_publication, 
    filename="emission_pathways_comparison_publication")

### Calculate R$^2$ scores of all the above regressions

# R2 metrics from the fitted EFs using linear regression
score_type = "flux"

metric_exec_dict_total = {
    "Soued": {
        "CO2 diffusion": ('co2_diffusion_total', 'co2_diffusion_total_soued'),
        "CH4 diffusion": ('ch4_diffusion_total', 'ch4_diffusion_total_soued'),
        "CH4 ebullition": ('ch4_ebullition_total', 'ch4_ebullition_total_soued'),
        "CH4 degassing": ('ch4_degassing_total', 'ch4_degassing_total_soued')
    },
    "Fitted": {
        "CO2 diffusion": ('co2_diffusion_total', 'co2_diffusion_total_fitted'),
        "CH4 diffusion": ('ch4_diffusion_total', 'ch4_diffusion_total_fitted'),
        "CH4 ebullition": ('ch4_ebullition_total', 'ch4_ebullition_total_fitted'),
        "CH4 degassing": ('ch4_degassing_total', 'ch4_degassing_total_fitted')        
    }
}
metric_exec_dict_flux = {
    "Soued": {
        "CO2 diffusion": ('co2_diffusion', 'co2_diffusion_soued'),
        "CH4 diffusion": ('ch4_diffusion', 'ch4_diffusion_soued'),
        "CH4 ebullition": ('ch4_ebullition', 'ch4_ebullition_soued'),
        "CH4 degassing": ('ch4_degassing', 'ch4_degassing_soued')
    },
    "Fitted": {
        "CO2 diffusion": ('co2_diffusion', 'co2_diffusion_fitted'),
        "CH4 diffusion": ('ch4_diffusion', 'ch4_diffusion_fitted'),
        "CH4 ebullition": ('ch4_ebullition', 'ch4_ebullition_fitted'),
        "CH4 degassing": ('ch4_degassing', 'ch4_degassing_fitted')        
    }
}

def calculate_r2_scores(df: pd.DataFrame, metric_exec_dict) -> pd.DataFrame:
    """ """
    r2_score_dict = {}
    for ef_option, pathway in metric_exec_dict.items():
        r2_score_dict[ef_option] = {}
        for pathway_name, mapping in pathway.items():
            r2_score_dict[ef_option][pathway_name] = r2_score(
                y_true=df[mapping[0]], y_pred = df[mapping[1]])
    return pd.DataFrame(r2_score_dict)

def calculate_mae_scores(df: pd.DataFrame, metric_exec_dict) -> pd.DataFrame:
    """ """
    mape_score_dict = {}
    for ef_option, pathway in metric_exec_dict.items():
        mape_score_dict[ef_option] = {}
        for pathway_name, mapping in pathway.items():
            mape_score_dict[ef_option][pathway_name] = mae(df[mapping[0]], df[mapping[1]])
    return pd.DataFrame(mape_score_dict)

# # Strategies: r_squared, rrmse, rmae, mape, rmse, mae

def calculate_score(df: pd.DataFrame, score: Callable, metric_exec_dict: Dict) -> pd.DataFrame:
    """ """
    score_dict = {}
    for ef_option, pathway in metric_exec_dict.items():
        score_dict[ef_option] = {}
        for pathway_name, mapping in pathway.items():
            score_dict[ef_option][pathway_name] = score(df[mapping[0]], df[mapping[1]])
    return pd.DataFrame(score_dict)    

    
r2s_flux = calculate_score(emissions_total, r2_score, metric_exec_dict_flux)
r2s_flux

r2s_total = calculate_score(emissions_total, r2_score, metric_exec_dict_total)
r2s_total

maes_flux = calculate_score(emissions_total, mae, metric_exec_dict_flux)
maes_flux

maes_total = calculate_score(emissions_total, mae, metric_exec_dict_total)
maes_total

rmses_total = calculate_score(emissions_total, rmse, metric_exec_dict_total)
rmses_total

rmses_flux = calculate_score(emissions_total, rmse, metric_exec_dict_flux)
rmses_flux

### Create a $\LaTeX$ table with fit quality metrics

flux_metrics_df = pd.concat([r2s_flux, maes_flux, rmses_flux], axis=1)
flux_metrics_df.columns = pd.MultiIndex.from_tuples(
    [("R2", col) for col in r2s_flux.columns] + [("MAE", col) for col in maes_flux.columns] +
    [("RMSE", col) for col in rmses_flux.columns],
    names=["Metric", "Option"]
)
metrics_latex_table = flux_metrics_df.to_latex(
    index=True, 
    float_format="%4.3f",
    caption="Levels of fit between emissions calculated with emission factors and the G-res method", 
    label="tab:emission_factors")
print(metrics_latex_table)

## Calculate and compare net emissions

net_emisions = (emissions_total.
     assign(
         em_net_total = lambda x: x['co2_net_total'] + x['ch4_net_total'],
         co2_net_total_soued=lambda x: x['co2_total_soued'] * net_CO2 * (1 + R_downstream),
         ch4_net_total_soued=lambda x: x['ch4_total_soued'] * net_CH4 * (1 + R_downstream),
         em_net_total_soued=lambda x: x['co2_net_total_soued'] + x['ch4_net_total_soued'],
         co2_net_total_fitted=lambda x: x['co2_total_fitted'] * net_CO2 * (1 + R_downstream),
         ch4_net_total_fitted=lambda x: x['ch4_total_fitted'] * net_CH4 * (1 + R_downstream),
         em_net_total_fitted=lambda x: x['co2_net_total_fitted'] + x['ch4_net_total_fitted']
     )
)

pairs_soued = [
    ('co2_net_total', 'co2_net_total_soued'), 
    ('ch4_net_total', 'ch4_net_total_soued'), 
    ('em_net_total', 'em_net_total_soued')
]
axis_labels_soued = {
    'co2_net_total' : 'Net total CO$_2$ emission (model), ktCO$_{2e}$/yr',
    'co2_net_total_soued': 'Net total CO$_2$ emission (EF), ktCO$_{2e}$/yr',
    'ch4_net_total': 'Net total CH$_4$ emission (model), ktCO$_{2e}$/yr',
    'ch4_net_total_soued': 'Net total CH$_4$ emission (EF), ktCO$_{2e}$/yr',
    'em_net_total': 'Net total emission (model), ktCO$_{2e}$/yr',
    'em_net_total_soued': 'Net total emission (EF), ktCO$_{2e}$/yr'}
pairs_fitted = [
    ('co2_net_total', 'co2_net_total_fitted'), 
    ('ch4_net_total', 'ch4_net_total_fitted'), 
    ('em_net_total', 'em_net_total_fitted')
]
axis_labels_fitted = {
    'co2_net_total' : 'Net total CO$_2$ emission (model), ktCO$_{2e}$/yr',
    'co2_net_total_fitted': 'Net total CO$_2$ emission (EF-fitted), ktCO$_{2e}$/yr',
    'ch4_net_total': 'Net total CH$_4$ emission (model), ktCO$_{2e}$/yr',
    'ch4_net_total_fitted': 'Net total CH$_4$ emission (EF-fitted), ktCO$_{2e}$/yr',
    'em_net_total': 'Net total emission (model), ktCO$_{2e}$/yr',
    'em_net_total_fitted': 'Net total emission (EF-fitted), ktCO$_{2e}$/yr'}

def make_pair_plots_total_emissions(
        df: pd.DataFrame, 
        pairs, axis_labels,
        file_suffix: str,
        dxdy1: float = 0.0, dxdy2: float = 300,
        save_fig: bool = save_figure_global) -> None:
    # Define the pairs of columns to plot
    sns.set_theme(style="ticks")
    sns.set_context("paper", font_scale=1.4) 
    label_fontsize = 10
    tick_fontsize = 15

    # Create the grid for subplots
    fig, axes = plt.subplots(nrows=1, ncols=3, figsize=(15, 5.0))
    axes = axes.flatten()
    # Generate q-q plots for each pair
    for idx, (col_x, col_y) in enumerate(pairs):
        ax = axes[idx]
        legend_status='auto'
        sns.scatterplot(
            x=col_x, y=col_y, style = 'type', hue='type', size='res_area', ax=ax, data=df,
            markers=['o', 's', '^'], alpha=0.5, sizes=(50, 500),
            **{'edgecolor':'black'})
        ax.set_xscale('log')
        ax.set_yscale('log')
        max_val = max(df[[col_x, col_y]].max())
        min_val = min(df[[col_x, col_y]].min())
        ax.plot([min_val, max_val], [min_val, max_val], 'k--', lw=3, alpha = 0.6)
        # Customize each subplot
        ax.set_xlabel(axis_labels[col_x])
        ax.set_ylabel(axis_labels[col_y])
        #ax.set_xticklabels(ax.get_xticks(), size = tick_fontsize)
        #ax.set_yticklabels(ax.get_yticks(), size = tick_fontsize)
        ax.set_ylim([min_val-dxdy1, max_val+dxdy2])
        ax.set_xlim([min_val-dxdy1, max_val+dxdy2])

    handles, labels = ax.get_legend_handles_labels()
    for handle in handles:
        if handle._label in ('hydroelectric', 'multipurpose', 'irrigation'):
            pass
            #handle.set_markersize(handle.get_markersize() * 2)
            
    fig.legend(
        handles, [label.replace('type', 'Type').replace('res_area', 'Area, km$^2$') for label in labels],
        loc='right', bbox_to_anchor=(1.085, 0.48), 
        scatterpoints=1, frameon=False)

    # Iterate again and remove legend from each subplot
    for idx, _ in enumerate(pairs):
        ax = axes[idx]
        ax.legend().remove()
    
    plt.tight_layout(rect=[0, 0, 0.95, 0.90])  # Adjust layout to make space for the legend
    fig.suptitle(
        "Comparison of Net Anthropogenic Emissions: Explicit Calculations vs. Emission Factor Estimates", 
        fontsize=18, fontweight='bold')
    if save_fig:
        fig.savefig(fig_folder / f'total emisssions_comparison_{file_suffix}.svg', bbox_inches='tight')
        fig.savefig(fig_folder / f'total emisssions_comparison_{file_suffix}.pdf', bbox_inches='tight')
    plt.show()

### Plot net emission fits using the default anthropogenic (net) contribution coefficients from Soued et al.

# Visualise net emissions fits based on Soued's Emission Factors
make_pair_plots_total_emissions(
    df=net_emisions, 
    pairs = pairs_soued, 
    axis_labels = axis_labels_soued,
    file_suffix="Soued")

# Visualise net emissions fits based on Fitted Emission Factors (see previous analysis)
make_pair_plots_total_emissions(
    df=net_emisions, 
    pairs = pairs_fitted, 
    axis_labels = axis_labels_fitted,
    file_suffix="Fitted")

## Calculate and plot errors between emission pathways and total emissions predicted from the model and using the emission factors (EF)

emissions_total.head()

comparison_pairs_soued = [
        ('co2_diffusion', 'co2_diffusion_soued'), 
        ('ch4_diffusion', 'ch4_diffusion_soued'),  
        ('ch4_ebullition', 'ch4_ebullition_soued'), 
        ('ch4_degassing', 'ch4_degassing_soued'),  
        ('co2', 'co2_soued'),
        ('ch4', 'ch4_soued')]

comparison_pairs_fitted = [
        ('co2_diffusion', 'co2_diffusion_fitted'), 
        ('ch4_diffusion', 'ch4_diffusion_fitted'),  
        ('ch4_ebullition', 'ch4_ebullition_fitted'), 
        ('ch4_degassing', 'ch4_degassing_fitted'),  
        ('co2', 'co2_fitted'),
        ('ch4', 'ch4_fitted')]

def create_hist_df(
        df: pd.DataFrame, 
        comparison_pairs,
        err_type: str) -> pd.DataFrame:
    emissions_long = pd.DataFrame({
        'g-res': pd.concat([df[model_value] for model_value, _ in comparison_pairs]),
        'ef': pd.concat([df[ef_value] for _, ef_value in comparison_pairs]),
        'type': pd.concat([df['type'] for _ in comparison_pairs]),
        'category': ['co2_diffusion'] * len(df) + 
                    ['ch4_diffusion'] * len(df) + 
                    ['ch4_ebullition'] * len(df) + 
                    ['ch4_degassing'] * len(df) + 
                    ['co2'] * len(df) + 
                    ['ch4'] * len(df),
        'err_type': [err_type] * len(df) * 6
    })
    # Calculate errors
    emissions_long['rel_err'] = \
        _rel_err(emissions_long['g-res'], emissions_long['ef'])
    emissions_long['rel_err_perc'] = \
        _rel_err(emissions_long['g-res'], emissions_long['ef']) * 100
    return emissions_long

emisions_long_soued = create_hist_df(
    df=emissions_total, 
    comparison_pairs = comparison_pairs_soued,
    err_type="Original")
emisions_long_fitted = create_hist_df(
    df=emissions_total, 
    comparison_pairs = comparison_pairs_fitted,
    err_type="Fitted")
emisions_long = pd.concat([emisions_long_soued, emisions_long_fitted])

emisions_long['rel_err'].describe()

emisions_long['rel_err_perc'].describe()

@skip_execution("This plot is replaced with a multifacet plot for each reservoir type.")
def make_box_plot_mape_all_dams(save_fig: bool = save_figure_global) -> None:
    # Visualise the errors on a box plot
    sns.set_theme(style="ticks")
    sns.set_context("paper", font_scale=1.6)
    # Initialize the figure with a logarithmic x axis
    f, ax = plt.subplots(figsize=(7, 6))
    sns.boxplot(
        emissions_long, y="mape", x="category", hue="category",
        whis=[0, 100], width=.6, palette="vlag"
    )
    sns.stripplot(emissions_long, y="mape", x="category", size=4, color=".3", alpha=0.6, edgecolor='black')
    # Tweak the visual presentation
    ax.xaxis.grid(False)
    ax.set(ylabel='Mean Absolute Percentage Error (MAPE)', xlabel="")
    ax.spines['left'].set_linewidth(0.7)
    ax.spines['bottom'].set_linewidth(0.7)
    
    # Adjust labels
    new_labels = [
        "CO$_2$ diffusion", "CH$_4$ diffusion", "CH$_4$ ebullition", 
        "CH$_4$ degassing", "CO$_2$ total", "CH$_4$ total"]
    ax.set_xticklabels(new_labels, rotation=45, ha="right")
    sns.despine(trim=True, left=False)
    if save_fig:
        f.savefig(fig_folder / 'mape_all_dams.svg', bbox_inches='tight')
        f.savefig(fig_folder / 'mape_all_dams.pdf', bbox_inches='tight')
    plt.show()
make_box_plot_mape_all_dams()

def make_box_plot_mape_multifaceted(
        df: pd.DataFrame = emissions_long, 
        y_val: str = "rel_err_perc",
        suptitle: bool = False,
        save_fig: bool = save_figure_global) -> None:
    # Visualise the errors on a box plot
    titles = {
        "hydroelectric": "Hydroelectric Reservoirs",
        "irrigation": "Irrigation Reservoirs",
        "multipurpose": "Multipurpose Reservoirs"
    }
    sns.set_theme(style="ticks")
    sns.set_context("paper", font_scale=1.8)
    # Initialize the figure with a logarithmic x axis
    g = sns.FacetGrid(df, col="type", height=6, aspect=1)
    g.map_dataframe(
        sns.boxplot, y=y_val, x="category", fliersize=0, hue="err_type",
        width=0.8, palette="vlag", linewidth=0.5
    )
    g.map_dataframe(
        sns.stripplot, y=y_val, x="category", size=3, color="lightgray", alpha=0.2, 
        edgecolor='black', dodge=True, linewidth=0.5, marker="o", hue="err_type",
        legend=False, palette=None
    )
    # Adjust labels
    new_labels = [
        "CO$_2$ diffusion", "CH$_4$ diffusion", "CH$_4$ ebullition", 
        "CH$_4$ degassing", "CO$_2$ total", "CH$_4$ total"]
    for ax in g.axes.flat:
        ax.xaxis.grid(False)
        ax.set_xlabel("")
        ax.set_ylabel("Relative Gross Emission Error, \%")
        ax.set_ylim((-250,1_250))
        ax.spines['left'].set_linewidth(0.7)
        ax.spines['bottom'].set_linewidth(0.7)
        ax.set_xticklabels(new_labels, rotation=45, ha="right")
    for ax, title in zip(g.axes.flat, g.col_names):
        ax.set_title(titles[title], fontsize=14, fontweight="bold", pad=25)  # Use the custom title dictionary
    plt.tight_layout(rect=[0, 0, 1, 0.93])  # Adjust layout to make space for the legend
    if suptitle:
        g.fig.suptitle(
            "Relative Errors in Reservoir Emission Predictions: Explicit Calculations vs. Emission Factor Estimates", 
            fontsize=18, fontweight='bold')
    g.add_legend(title="Emission Factors")
        
    g.despine(trim=True, left=False)
    if save_fig:
        g.fig.savefig(fig_folder / 'mape_all_dams_faceted.svg', bbox_inches='tight')
        g.fig.savefig(fig_folder / 'mape_all_dams_faceted.pdf', bbox_inches='tight')
    plt.show()
    
make_box_plot_mape_multifaceted(emisions_long)

emissions_long

emissions_long.query("type == 'hydroelectric' & category == 'ch4_ebullition'").describe(
    percentiles=[.9, .95, .997])

## Calculate and plot errors between net emissions from the model and the emission factors (EF)

report_errors(net_emisions['co2_net_total'], net_emisions['co2_net_total_soued'])

report_errors(net_emisions['ch4_net_total'], net_emisions['ch4_net_total_soued'])

report_errors(net_emisions['em_net_total'], net_emisions['em_net_total_soued'])

net_emisions['em_net_mape'] = \
    _rel_err2(
        y_true = net_emisions['em_net_total'], 
        y_pred = net_emisions['em_net_total_soued']) * 100 * -1

def make_mape_histogram(save_fig: bool = save_figure_global) -> None:
    """ """
    hydro = net_emisions["em_net_mape"].where(net_emisions['type']=='hydroelectric').dropna()
    irr = net_emisions["em_net_mape"].where(net_emisions['type']=='irrigation').dropna()
    multi = net_emisions["em_net_mape"].where(net_emisions['type']=='multipurpose').dropna()
    label_fontsize = 15
    tick_fontsize = 14
    plt.style.use('seaborn-v0_8-ticks')
    f, ax = plt.subplots(figsize=(7, 5.5))
    ax.tick_params(axis='both', which='major', labelsize=tick_fontsize)
    min_val = net_emisions['em_net_mape'].min()
    max_val = net_emisions['em_net_mape'].max()
    bins = np.linspace(int(min_val), int(max_val), 18)
    plt.xlabel('Mean Absolute Percentage Error (MAPE)', fontsize=label_fontsize)
    plt.ylabel('Frequency', fontsize=label_fontsize)
    plt.hist(
        [hydro, irr, multi], 
        label=['hydro', 'irrigation', 'multipurpose'], 
        histtype='bar', bins=bins, edgecolor='black', alpha=0.7)
    plt.legend(loc='upper right', frameon=False, fontsize=14)
    # Remove ticks from top and right side
    ax.tick_params(top=False, right=False)
    f.tight_layout(rect=[0, 0, 1, 0.90])  # Adjust layout to make space for the legend
    f.suptitle(
        "Relative Errors in Net Anthropogenic Emissions:\nExplicit Calculations vs. Emission Factor Estimates",
        fontweight='bold')
    # Decrease the width of the plot border
    for spine in ax.spines.values():
        spine.set_linewidth(0.6)
    if save_fig:
        f.savefig(fig_folder / 'mape_all_dams_histogram.svg', bbox_inches='tight')
        f.savefig(fig_folder / 'mape_all_dams_histogram.pdf', bbox_inches='tight')
    plt.show()
make_mape_histogram()

### Run Linear Regression to calibrate net$_{CO_2}$ and net$_{CH_4}$ parameters that are used to estimate net anthropogenic emissions from total emissions

* **NOTE**: Differences between explicitly calculated emissions and emission factors originate from two sources:
* 1. Emission factors are binned into categories and are therefore constant over an area of interest e.g. latitude band or climatic zone, whilst in reality, emissions vary continuously in response to many continuously changing emission drivers.
  2. Disentanglement of anthropogenic emissions from total, e.g. measured emissions is a function of complex relationships such as pre-impoundment emissions, levels of landuse, impacts of nutrients originating from human activities, whilst in case of applying simple emission factors, a single regression is applied individuall to CO$_2$ and CH$_4$ emissions. This simplifies the analysis but introduces a new source of error.

Here, we fit these regressions to our emission estimates and show that the obtained parameters substantially differ from the coefficient values adopted in previous studies, e.g. Almeida et al. 2019 [[3](https://doi.org/10.1038/s41467-019-12179-5)] and Carlino et al. 2024 [[4](https://doi.org/10.1038/s41893-024-01367-x)] 

def find_confidence_intervals_net_regression(
        X: pd.DataFrame,
        y: pd.Series,
        model,
        n_bootstraps: int = 1_000, 
        confidence_level: float = 0.95) -> pd.Series:
    """ """
    # Store bootstrap coefficients
    bootstrap_coefficients = []
    # Perform bootstrapping
    for i in range(n_bootstraps):
        model_cf = clone(model)
        # Resample data with replacement
        X_resampled, y_resampled = resample(X, y)
        model_cf.fit(X_resampled, y_resampled)
        base_coefficients = model_cf.coef_
        # Store the coefficients
        bootstrap_coefficients.append(base_coefficients)
        
    # Convert bootstrap coefficients to a DataFrame
    bootstrap_coefficients = np.array(bootstrap_coefficients)
    bootstrap_coefficients_2d = np.squeeze(bootstrap_coefficients, axis=(1, 2))
    bootstrap_df = pd.DataFrame(bootstrap_coefficients_2d, columns=X.columns)
    # Calculate the confidence intervals for each coefficient
    ci_lower = bootstrap_df.quantile((1 - confidence_level) / 2)
    ci_upper = bootstrap_df.quantile(1 - (1 - confidence_level) / 2)
    deltas = (ci_upper - ci_lower) / 2
    return deltas

net_em_lr_results = {
    "net_CO2": None,
    "net_CO2_CI": None,
    "net_CO2_R2": None,
    "net_CO2_MAE": None,
    "net_CO2_RMSE": None,
    "net_CH4": None,
    "net_CH4_CI": None,
    "net_CH4_R2": None,
    "net_CH4_MAE": None,
    "net_CH4_RMSE": None,
    "Rdownstream": 0.17
}

### 1. Fit the model using EF derived emissions with default values from Soued et al.

# CO2 net emissions
co2_net_total = net_emisions[['co2_net_total']] # Net total from G-res
co2_total_soued = net_emisions[['co2_total_soued']] # Net total from Soued's emission factors
co2_total_fitted = net_emisions[['co2_total_fitted']] # Net total from fitted emission factors
# CH4 net emissions
ch4_net_total = net_emisions[['ch4_net_total']] # Net total from G-res
ch4_total_soued = net_emisions[['ch4_total_soued']] # Net total from Soued's emission factors
ch4_total_fitted = net_emisions[['ch4_total_fitted']] # Net total from fitted emission factors

def fit_net_emission_model(
        co2, 
        co2_net, 
        ch4, 
        ch4_net,
        R_downstream = 0.17) -> Dict:
    """ """
    # Set up regression model objects
    model_co2 = LinearRegression(fit_intercept=False)
    model_ch4 = LinearRegression(fit_intercept=False)
    # Fit models
    model_co2.fit(co2, co2_net)
    model_ch4.fit(ch4, ch4_net)
    # Get fitted parameters
    m_co2 = model_co2.coef_[0]
    m_ch4 = model_ch4.coef_[0]
    net_CO2_fitted = m_co2 / (1+R_downstream)
    net_CH4_fitted = m_ch4 / (1+R_downstream)
    # Find confidence intervals on the fitted parameters
    ci_co2 = find_confidence_intervals_net_regression(co2, co2_net, model_co2)
    ci_ch4 = find_confidence_intervals_net_regression(ch4, ch4_net, model_ch4)
    # Predict net emissions from the regressions
    co2_net_lr = net_CO2_fitted * co2 * (1+R_downstream)
    ch4_net_lr = net_CH4_fitted * ch4 * (1+R_downstream)
    
    r2_score_co2 = r_squared(co2_net_lr, co2_net)
    mae_score_co2 = mae(co2_net_lr.values, co2_net.values)
    rmse_score_co2 = rmse(co2_net_lr.values, co2_net.values)
    r2_score_ch4 = r_squared(ch4_net_lr, ch4_net)
    mae_score_ch4 = mae(ch4_net_lr.values, ch4_net.values)
    rmse_score_ch4 = rmse(ch4_net_lr.values, ch4_net.values)
    out = {
        'net_CO2': net_CO2_fitted[0],
        'net_CO2_CI': ci_co2[0],
        'net_CO2_R2': r2_score_co2,
        'net_CO2_MAE': mae_score_co2,
        'net_CO2_RMSE': rmse_score_co2,
        'net_CH4': net_CH4_fitted[0],
        'net_CH4_CI': ci_ch4[0],
        'net_CH4_R2': r2_score_ch4,
        'net_CH4_MAE': mae_score_ch4,
        'net_CH4_RMSE': rmse_score_ch4,
    }
    return out
    
outputs_lr_soued = fit_net_emission_model(
    co2 = co2_total_soued, 
    co2_net = co2_net_total,
    ch4 = ch4_total_soued, 
    ch4_net = ch4_net_total)
outputs_lr_soued

outputs_lr_fitted = fit_net_emission_model(
    co2 = co2_total_fitted, 
    co2_net = co2_net_total,
    ch4 = ch4_total_fitted, 
    ch4_net = ch4_net_total)
outputs_lr_fitted

def calc_net_co2(emission_co2, net_CO2 = 0.25, R_downstream = 0.17):
    return emission_co2 * net_CO2 * (1+R_downstream)
def calc_net_ch4(emission_ch4, net_CH4 = 0.9, R_downstream = 0.17):
    return emission_ch4 * net_CH4 * (1+R_downstream)

# Calculate metrics with the original (default coefficients)
outputs_default = {
    "net_CO2": 0.25,
    "net_CO2_CI": None,
    "net_CO2_R2": r_squared(calc_net_co2(co2_total_soued), co2_net_total),
    'net_CO2_MAE': mae(calc_net_co2(co2_total_soued).values, co2_net_total.values),
    'net_CO2_RMSE': rmse(calc_net_co2(co2_total_soued).values, co2_net_total.values),
    'net_CH4': 0.90,
    'net_CH4_CI': None,
    'net_CH4_R2': r_squared(calc_net_ch4(ch4_total_soued), ch4_net_total),
    'net_CH4_MAE': mae(calc_net_ch4(ch4_total_soued).values, ch4_net_total.values),
    'net_CH4_RMSE': rmse(calc_net_ch4(ch4_total_soued).values, ch4_net_total.values)
}
outputs_default

R_downstream = 0.17
net_emissions = (net_emisions.
     assign(
         em_net_total = lambda x: x['co2_net_total'] + x['ch4_net_total'],
         co2_net_total_soued_fitted=lambda x: x['co2_total_soued'] * outputs_lr_soued['net_CO2'] * (1 + R_downstream),
         ch4_net_total_soued_fitted=lambda x: x['ch4_total_soued'] * outputs_lr_soued['net_CH4'] * (1 + R_downstream),
         em_net_total_soued_fitted=lambda x: x['co2_net_total_soued_fitted'] + x['ch4_net_total_soued_fitted'],
         co2_net_total_fitted_fitted=lambda x: x['co2_total_soued'] * outputs_lr_fitted['net_CO2'] * (1 + R_downstream),
         ch4_net_total_fitted_fitted=lambda x: x['ch4_total_soued'] * outputs_lr_fitted['net_CH4'] * (1 + R_downstream),
         em_net_total_fitted_fitted=lambda x: x['co2_net_total_fitted_fitted'] + x['ch4_net_total_fitted_fitted']
     )
)

net_emissions.columns

pairs_net_emissions = [
    ('co2_net_total', 'co2_net_total_soued'), 
    ('ch4_net_total', 'ch4_net_total_soued'), 
    ('em_net_total', 'em_net_total_soued'),
    ('co2_net_total', 'co2_net_total_soued_fitted'), 
    ('ch4_net_total', 'ch4_net_total_soued_fitted'), 
    ('em_net_total', 'em_net_total_soued_fitted')
]
axis_labels_net_emissions = {
    'co2_net_total' : 'Net total CO$_2$ emission (model), ktCO$_{2e}$/yr',
    'co2_net_total_soued': 'Net total CO$_2$ emission (EF), ktCO$_{2e}$/yr',
    'co2_net_total_soued_fitted': 'Net total CO$_2$ emission (EF) - regression, ktCO$_{2e}$/yr',
    'ch4_net_total': 'Net total CH$_4$ emission (model), ktCO$_{2e}$/yr',
    'ch4_net_total_soued': 'Net total CH$_4$ emission (EF), ktCO$_{2e}$/yr',
    'ch4_net_total_soued_fitted': 'Net total CH$_4$ emission (EF) - regression, ktCO$_{2e}$/yr',
    'em_net_total': 'Net total emission (model), ktCO$_{2e}$/yr',
    'em_net_total_soued': 'Net total emission (EF), ktCO$_{2e}$/yr',
    'em_net_total_soued_fitted': 'Net total emission (EF) - regression, ktCO$_{2e}$/yr'}

def make_net_emission_regression_plots(
        df: pd.DataFrame, 
        pairs, axis_labels,
        file_suffix: str,
        dxdy1: float = 0.0, dxdy2: float = 300,
        save_fig: bool = save_figure_global) -> None:
    # Define the pairs of columns to plot
    sns.set_theme(style="ticks")
    sns.set_context("paper", font_scale=1.4) 
    label_fontsize = 10
    tick_fontsize = 15

    # Create the grid for subplots
    fig, axes = plt.subplots(nrows=2, ncols=3, figsize=(15, 10.0))
    axes = axes.flatten()
    # Generate q-q plots for each pair
    for idx, (col_x, col_y) in enumerate(pairs):
        ax = axes[idx]
        legend_status='auto'
        sns.scatterplot(
            x=col_x, y=col_y, style = 'type', hue='type', size='res_area', ax=ax, data=df,
            markers=['o', 's', '^'], alpha=0.5, sizes=(50, 500),
            **{'edgecolor':'black'})
        ax.set_xscale('log')
        ax.set_yscale('log')
        max_val = max(df[[col_x, col_y]].max())
        min_val = min(df[[col_x, col_y]].min())
        ax.plot([min_val, max_val], [min_val, max_val], 'k--', lw=3, alpha = 0.6)
        # Customize each subplot
        ax.set_xlabel(axis_labels[col_x])
        ax.set_ylabel(axis_labels[col_y])
        #ax.set_xticklabels(ax.get_xticks(), size = tick_fontsize)
        #ax.set_yticklabels(ax.get_yticks(), size = tick_fontsize)
        ax.set_ylim([min_val-dxdy1, max_val+dxdy2])
        ax.set_xlim([min_val-dxdy1, max_val+dxdy2])

    handles, labels = ax.get_legend_handles_labels()
    for handle in handles:
        if handle._label in ('hydroelectric', 'multipurpose', 'irrigation'):
            pass
            #handle.set_markersize(handle.get_markersize() * 2)
            
    fig.legend(
        handles, [label.replace('type', 'Type').replace('res_area', 'Area, km$^2$') for label in labels],
        loc='right', bbox_to_anchor=(1.085, 0.48), 
        scatterpoints=1, frameon=False)

    # Iterate again and remove legend from each subplot
    for idx, _ in enumerate(pairs):
        ax = axes[idx]
        ax.legend().remove()
    
    plt.tight_layout(rect=[0, 0, 0.95, 0.90])  # Adjust layout to make space for the legend
    fig.suptitle(
        "Comparison of Net Anthropogenic Emissions: Explicit Calculations vs. Emission Factor Estimates", 
        fontsize=18, fontweight='bold')
    if save_fig:
        fig.savefig(fig_folder / f'net emisssions_comparison_{file_suffix}.svg', bbox_inches='tight')
        fig.savefig(fig_folder / f'net emisssions_comparison_{file_suffix}.pdf', bbox_inches='tight')
    plt.show()

make_net_emission_regression_plots(
    df=net_emissions,
    pairs=pairs_net_emissions, 
    axis_labels=axis_labels_net_emissions,
    file_suffix="soued")

outputs_lr_fitted
outputs_lr_soued
outputs_default

outputs_lr_soued

# Prepare parameter value and fit metrics tables
table1_dict = {
    "original_value": [0.25, 0.90],
    "original_R2": [
        outputs_default['net_CO2_R2'],
        outputs_default['net_CH4_R2']
    ],
    "original_MAE": [
        outputs_default['net_CO2_MAE'],
        outputs_default['net_CH4_MAE']
    ],
    "original_RMSE": [
        outputs_default['net_CO2_RMSE'],
        outputs_default['net_CH4_RMSE']
    ],    
    "EF": [
        outputs_lr_soued['net_CO2'], 
        outputs_lr_soued['net_CH4']
    ],
    "EF_R2": [
        outputs_lr_soued['net_CO2_R2'], 
        outputs_lr_soued['net_CH4_R2']
    ],
    "EF_MAE": [
        outputs_lr_soued['net_CO2_MAE'], 
        outputs_lr_soued['net_CH4_MAE']
    ],
    "EF_RMSE": [
        outputs_lr_soued['net_CO2_RMSE'], 
        outputs_lr_soued['net_CH4_RMSE']
    ],  
    "EF-fitted": [
        outputs_lr_fitted['net_CO2'], 
        outputs_lr_fitted['net_CH4']
    ],
    "EF-fitted_R2": [
        outputs_lr_fitted['net_CO2_R2'], 
        outputs_lr_fitted['net_CH4_R2']
    ],
    "EF-fitted_MAE": [
        outputs_lr_fitted['net_CO2_MAE'], 
        outputs_lr_fitted['net_CH4_MAE']
    ],
    "EF-fitted_RMSE": [
        outputs_lr_fitted['net_CO2_RMSE'], 
        outputs_lr_fitted['net_CH4_RMSE']
    ],  
}
table1 = pd.DataFrame(table1_dict, index = ["net_CO2", "net_CH4"])
# Create a latex table with regression coefficients 
table1_latex = table1.to_latex(
    index=True, 
    float_format="%4.3f",
    caption="Net emission regression coefficients", 
    label="tab:net_emission_parameters")
print(table1_latex)

net_emissions['em_net_mape_fitted'] = \
    _rel_err2(
        y_true = net_emissions['em_net_total'], 
        y_pred = net_emissions['em_net_total_fitted']) * 100 * -1
net_emissions['em_net_mape_fitted_fitted'] = \
    _rel_err2(
        y_true = net_emissions['em_net_total'], 
        y_pred = net_emissions['em_net_total_fitted_fitted']) * 100 * -1

# Prepare data for plotting
net_emissions_long_rel_net_em = pd.melt(
    net_emissions,
    id_vars=["type"],
    value_vars=["em_net_mape", "em_net_mape_fitted", "em_net_mape_fitted_fitted"],
    var_name="error_type",
    value_name="error_value"
)

net_emissions_long_rel_net_em

# Visualise the errors on a box plot
def make_box_plot_mape_net_emissions(
        df: pd.DataFrame = net_emissions_long_rel_net_em,
        box_width: float = 0.8,
        suptitle: bool = False,
        save_fig: bool = save_figure_global) -> None:
    """ """
    df = df.copy()
    df["error_type"] = df["error_type"].replace({
        "em_net_mape": "Default EF and net regression coefficients",
        "em_net_mape_fitted": "Fitted EF and default net regression coefficients",
        "em_net_mape_fitted_fitted": "Fitted EF and net regression coefficients"
    })
    sns.set_theme(style="ticks")
    sns.set_context("paper", font_scale=1.6)
    # Initialize the figure with a logarithmic x axis
    f, ax = plt.subplots(figsize=(7, 6))
    # Plot the orbital period with horizontal boxes
    sns.boxplot(
        df, y="error_value", x="type", hue="error_type",
        showfliers = False, linewidth=0.5,
        width=box_width, palette="vlag", ax=ax
    )
    
    ## Add annotations
    for patch in ax.artists:
        r, g, b, a = patch.get_facecolor()
        patch.set_facecolor((r, g, b, .3))
        x = patch.get_x()
        y = patch.get_y()
        width = patch.get_width()
        height = patch.get_height()
        ax.text(x+width/2, y+height/2, "{:.2f}".format(height), ha='center', va='center')
    
    # Add in points to show each observation
    sns.stripplot(
        df, y="error_value", x="type", hue="error_type", size=5, color="lightgray", 
        ax = ax,
        alpha=0.2, dodge=True,
        edgecolor='black', linewidth=0.5)  # Subtle variation for stripplot points)
    
    # Add a legend
    handles, labels = ax.get_legend_handles_labels()
    ax.legend(
        handles[:3], labels[:3], loc="upper left", title=None,
        bbox_to_anchor=(0.01, 0.91),
        fontsize="small")
    
    # Tweak the visual presentation
    f.tight_layout(rect=[0, 0, 1, 0.90])  # Adjust layout to make space for the legend
    if suptitle:
        f.suptitle(
            "Relative Errors in Net Anthropogenic Emissions:\nExplicit Calculations vs. Emission Factor Estimates",
            fontweight='bold')
    ax.xaxis.grid(False)
    ax.set(ylabel='Relative Net Emission Error, %', xlabel="")
    ax.spines['left'].set_linewidth(0.7)
    ax.spines['bottom'].set_linewidth(0.7)
    sns.despine(trim=True, left=False, offset = 5)
    if save_fig:
        f.savefig(fig_folder / 'mape_all_dams_histogram_net_emissions.svg', bbox_inches='tight')
        f.savefig(fig_folder / 'mape_all_dams_histogram_net_emissions.pdf', bbox_inches='tight')
    plt.show()
make_box_plot_mape_net_emissions(box_width = 0.8, save_fig=True)
